==> src/tier_1/1 - Create Vocabulary and Pronunciations.ipynb <==
%load_ext rpy2.ipython
import rpy2.robjects.lib.ggplot2 as ggplot2
import childespy
import numpy as np
import os
import imp
import pandas as pd
import Levenshtein
import itertools
import re

import sys
sys.path.append('../../')
from src.utils import configuration, phonology
config = configuration.Config()

# Extract IPA Annotated Words from Providence

pvd_idx = childespy.get_sql_query('select * from corpus where name = "Providence"',
                                 db_version = "2020.1").iloc[0]['id']

pvd_chi_tokens = childespy.get_sql_query('select gloss, target_child_name, target_child_age, \
speaker_code, actual_phonology, model_phonology, transcript_id, utterance_id, \
token_order from token where speaker_code = "CHI" and corpus_id = '+str(pvd_idx),
    db_version = "2020.1")
pvd_chi_tokens.to_csv(os.path.join(config.project_root,'output/csv/pvd_tokens.csv'), index=False)


excludes = ['*','(.)','(..)', '(...)','(....)','(.....)'] # these are either unlabeled, or marked for number of syllables (but not for phonology)
pvd_chi_tokens = pvd_chi_tokens.loc[~(pvd_chi_tokens.model_phonology.isin(excludes) |
    pvd_chi_tokens.actual_phonology.isin(excludes))]

pvd_chi_tokens.iloc[0]

# get the most common transcription for each form
pvd_chi_tokens.gloss = [x.lower() for x in pvd_chi_tokens.gloss]
ipa_for_glosses = pvd_chi_tokens.groupby(['gloss']).model_phonology.agg(lambda x:
    x.value_counts().reset_index().iloc[0]['index']    
).reset_index()

print(ipa_for_glosses.shape)
test = ipa_for_glosses.dropna()
test.shape

test_words = ['dog','cat','box','fishes', 'mommy', 'potato','a', 'read']
ipa_for_glosses.loc[ipa_for_glosses.gloss.isin(test_words)]

import imp
imp.reload(phonology)

ipa_for_glosses = ipa_for_glosses.loc[ipa_for_glosses.model_phonology != ""]
ipa_for_glosses['model_phonology'] = [phonology.strip_accents(phonology.phone_remap(x)) for x in ipa_for_glosses['model_phonology']]
ipa_for_glosses.shape


ipa_for_glosses.iloc[0:3]

Using just the model phonology from Providence is insufficient. So we map to citation forms in the CMU pronouncing dictionary.

# Link to CMU Citation Forms

cmu = pd.read_csv(os.path.join(config.project_root,'data/cmu_dict_df.csv'), encoding='latin1') # load an unedited version of the CMU pronunciation dictionary
cmu.columns = ['index','word','pronunciation']
from string import digits
remove_digits = str.maketrans('', '', digits)
cmu['phones'] = [[y.translate(remove_digits) for y in x.split(' ')] for x in cmu['pronunciation']] # remove indicators of stress marking 
cmu.word = [str(x).lower() for x in cmu.word]

# Output a CSV with all phonemes in the CMU (phon/phon_map.csv). Manually map these to phonemes and then load it (phon/phon_map_populated.csv)

# def flatten(list_of_lists):
#     return([item for subl in list_of_lists for item in subl])

# phone_inventory = pd.DataFrame({'arpa':np.unique(flatten(cmu.phones))})
# phone_inventory.to_csv('phon/phon_map.csv', index=False)


# Character differences / shorthand
# [X] cmu: ɝ -> əɹ
# [X] cmu:  tʃ -> ʧ
# [X] cmu  dʒ -> ʤ      
# Collapses
# [X] cmu  I -> schwa
# [X] ɑ in model phonlogy vs ɔ in cmu: convert ɔ to ɑ in both
# [X] model phonology: ʌ ->  ə

phone_map_df = pd.read_csv(os.path.join(config.project_root,'data/phon_map_populated.csv'))
phone_map = dict(zip(phone_map_df.arpa, phone_map_df.ipa))
cv_map = dict(zip(phone_map_df.arpa, phone_map_df.c_or_v))
cmu['ipa'] = [[phone_map[x] for x in y] for y in cmu.phones]
cmu['structure'] = [[cv_map[x] for x in y] for y in cmu.phones]
cmu['num_vowels'] = [np.sum(np.array(x) == 'v') for x in cmu['structure']]
cmu['ipa_short'] = [''.join(x) for x in cmu['ipa']]
cmu['ipa_short'] =  [x.replace('ɝ', 'əɹ').replace('ɪ','ə').replace('ɔ','ɑ').replace('ər', 'əɹ').replace('er','əɹ').replace('ʧ','tʃ').replace('ʤ','dʒ') for x in cmu['ipa_short']]
# remove any parenthetical + contents from the word record in CMU
cmu['word'] = [re.sub("\(.+\)$", '', x) for x in cmu['word']]

# confirm that we have multiple pronunciation entries for a word like "read"
cmu.loc[cmu.word=="read"]

cmu.loc[cmu.word.isin(test_words)]

## Compute Proportion of Providence words that are in the adjusted CMU 

found_words = ipa_for_glosses.merge(cmu, left_on=['gloss'],
    right_on = ['word'])
found_words.drop_duplicates(subset=['word'])
# drop duplicates for a given word

found_words.shape[0] / ipa_for_glosses.shape[0]

Approximately 95% of the words in Providence are also present in CMU

# different pronunciations
missing_pronunctiations = ipa_for_glosses.loc[~ipa_for_glosses.gloss.isin(found_words.gloss)]
missing_pronunctiations_short = missing_pronunctiations[['gloss', 'model_phonology']]
missing_pronunctiations_short.sample(10)

These are justifiably odd words

## Compute Proportion of Providence annotations that match the annotation in adjusted CMU 

matched_words = ipa_for_glosses.merge(cmu, left_on=['gloss','model_phonology'],
    right_on = ['word','ipa_short'])
matched_words.shape[0] / ipa_for_glosses.shape[0]

Approximately 86\% of words have the same pronunciation in the corrected CMU as in Providence

# different pronunciations
different_pronunctiations = ipa_for_glosses.loc[~ipa_for_glosses.gloss.isin(matched_words.gloss)]
different_pronunctiations = different_pronunctiations.merge(cmu, left_on=['gloss'], right_on=['word'])
different_pronunctiations_short = different_pronunctiations[['gloss', 'model_phonology','ipa_short']]
different_pronunctiations_short.columns = ['gloss','pvd_ipa','cmu_ipa']
different_pronunctiations_short.sample(10)

childes_counts = pd.read_csv(os.path.join(config.project_root, 'data/vocab.csv'))
childes_counts.columns 
cmu_in_childes = cmu.loc[(cmu.word.isin(
    childes_counts.loc[childes_counts['count'] > 3].word)) & (cmu['num_vowels'] <=2)]

cmu_in_childes.shape

# Compare with Coverage in the BERT Tokenizer

# there's a circularity here -- we need a BERT model in order to know what all is in the tokenizer

from transformers import BertTokenizer
initial_tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')


bert_tokenizer_vocab = initial_tokenizer.get_vocab().keys()

cmu_in_childes.shape

cmu_in_childes_with_tokenizer = cmu.loc[(cmu.word.isin(
    childes_counts.loc[childes_counts['count'] > 3].word)) & (cmu['num_vowels'] <=2) &
            (cmu.word.isin(bert_tokenizer_vocab))]
cmu_in_childes_with_tokenizer.shape

cmu_in_childes_with_tokenizer.to_csv(os.path.join(config.project_root,'data/cmu_in_childes.csv'),index=False)
cmu_in_childes_with_tokenizer.to_pickle(os.path.join(config.project_root,'output/pkl/cmu_in_childes.pkl'))


cmu_in_childes_with_tokenizer.loc[cmu_in_childes_with_tokenizer.word =='read']

cmu_in_childes_with_tokenizer.loc[cmu_in_childes_with_tokenizer.word == 'watch']

# Compute Coverage of MacArthur CDI words

wswg  = pd.read_csv(os.path.join(config.project_root, 'data/WSWG_50percentproducing.csv')) # load a list of all Words and Sentences + Words and Gestures words

wswg['word'] = [x.split(' ')[0].split('/')[0] for x in wswg.word]
' '.join(wswg['word'])
wswg_cmu = wswg.merge(cmu_in_childes)
wswg_cmu = wswg_cmu.drop_duplicates(subset=['word'])
print('Total coverage of MacArthur CDI words, before intersecting with BERT tokenizer')
print(wswg_cmu.shape[0] / wswg.shape[0])


wswg['word'] = [x.split(' ')[0].split('/')[0] for x in wswg.word]
' '.join(wswg['word'])
wswg_cmu = wswg.merge(cmu_in_childes_with_tokenizer)
wswg_cmu = wswg_cmu.drop_duplicates(subset=['word'])
print('Total coverage of MacArthur CDI words, after intersecting with BERT tokenizer')
print(wswg_cmu.shape[0] / wswg.shape[0])

# What words from the CDI are missing? 
' '.join(wswg.loc[~wswg.word.isin(wswg_cmu.word)].word)

We are missing multisyllable words, onomatopoeia from CHILDES
