==> src/tier_1/3 - Providence - Retrieve Data.ipynb <==
%load_ext rpy2.ipython
import rpy2.robjects.lib.ggplot2 as ggplot2
import childespy
import numpy as np
import os
import sys
import imp
import pandas as pd
import transformers
import torch
import re
import unicodedata
import scipy.stats
import copy
from string import punctuation
from os.path import join, exists

sys.path.append('../../')
from src.utils import configuration, load_models, transformers_bert_completions, data_cleaning, phonology
config = configuration.Config()

# Identify "Communicative Successes" and "Communicative Failures" 

Defintions:  
`xxx`: no lexical label, phonetically un-transcribable  
`yyy`: no lexical label, phonetically transcribable  
`Communicative success`: token from utterances with no xxx's or yyy's   
`Communicative failure`: a yyy token from an utterance with no other xxx's or yyy's

Later, we'll further restrict both to tokens that are monosyllabic

pvd_idx = childespy.get_sql_query('select * from corpus where name = "Providence"', db_version = "2020.1").iloc[0]['id']

phono_glosses = childespy.get_sql_query('select gloss, target_child_name, target_child_age, \
    speaker_code, actual_phonology, model_phonology, transcript_id, utterance_id, \
    token_order, corpus_name, collection_name, language from token where \
    actual_phonology != "" and model_phonology != "" and speaker_code in ("MOT", "FAT","CHI") and collection_name = "Eng-NA" \
    and corpus_id = '+str(pvd_idx) ,
        db_version = "2020.1")

phono_glosses.corpus_name.value_counts()

phono_glosses.loc[phono_glosses.gloss == 'xxx'].actual_phonology.value_counts()

Actual phonology is almost always * for xxx items

phono_glosses.loc[phono_glosses.gloss == 'yyy'].actual_phonology.value_counts()

Actual phonology *is* populated for tokens with `yyy` lexical label

chi_phono = phono_glosses.loc[(phono_glosses.speaker_code == 'CHI') & 
    (phono_glosses.target_child_age < (365*5))]

def count_transmission_errors(utt_vector, error_codes):
    return(np.sum([x in error_codes for x in  utt_vector]))

xxxs_per_utt = chi_phono.groupby('utterance_id').gloss.agg(
    lambda x: count_transmission_errors(x, ['xxx'])).reset_index()
xxxs_per_utt.columns = ['utterance_id', 'num_xxx']
yyys_per_utt = chi_phono.groupby('utterance_id').gloss.agg(
    lambda x: count_transmission_errors(x, ['yyy'])).reset_index()
yyys_per_utt.columns = ['utterance_id', 'num_yyy']
failures_per_utt = xxxs_per_utt.merge(yyys_per_utt)

yyy_utts = failures_per_utt.loc[(failures_per_utt.num_xxx == 0) &  (failures_per_utt.num_yyy == 1)]

yyy_utts.shape

success_utts = failures_per_utt.loc[(failures_per_utt.num_xxx == 0) &  
    (failures_per_utt.num_yyy == 0)]

success_utts.shape

tokens_from_errorless_utts = chi_phono.loc[chi_phono.utterance_id.isin(success_utts.utterance_id)]

#exclude un-transcribed tokens and syllabically transcribed tokens
excludes = ['*','(.)','(..)', '(...)','(....)','(.....)']
tokens_from_errorless_utts = tokens_from_errorless_utts.loc[~(tokens_from_errorless_utts.actual_phonology.isin(excludes) |
    tokens_from_errorless_utts.model_phonology.isin(excludes))]

tokens_from_errorless_utts.shape

# example phonology
tokens_from_errorless_utts.actual_phonology

# 31,457 transmission errors (from 31,457 utterances)
# 214,239 transmission successes (from 83,880 utterances)
# this will be further decreased later by the need to test monosyllabic forms

# Load BERT Models + CMU Dict

### Get the CMU Pronunciation Dictionary

import imp
imp.reload(configuration)
config = configuration.Config()

cmu_in_childes = pd.read_pickle(config.cmu_path)
cmu_2syl_inchildes = cmu_in_childes.loc[cmu_in_childes.num_vowels <=2]

cmu_in_childes.shape

# Prep Utterances / Tokens for BERT

# Load utterances from the Providence corpus from childs-db

this_path = join(config.prov_csv_dir, 'pvd_utt_glosses.csv')
if config.regenerate:
    utt_glosses = childespy.get_sql_query('select gloss, transcript_id, id, \
    utterance_order, target_child_name, speaker_code, type from utterance where speaker_code in ("MOT", "FAT","CHI") and corpus_id = '+str(pvd_idx) ,
        db_version = "2020.1")
    utt_glosses.to_csv(this_path, index=False)
else: 
    utt_glosses = pd.read_csv(this_path)

utt_glosses = utt_glosses.rename(columns = {'id' : 'utterance_id'})

# Prep the utterances for tokenization with BERT. Importantly, add back the appropriate punctutation for the sentence type
utt_glosses = utt_glosses[~utt_glosses.gloss.isna()]
utt_glosses = data_cleaning.clean_glosses_and_add_speaker_label(utt_glosses, True)

utt_glosses.head(5)

### Build the Vocabulary

import imp
imp.reload(load_models)
imp.reload(transformers_bert_completions)

initial_tokenizer = load_models.get_primary_tokenizer()
initial_vocab, cmu_in_initial_vocab, cmu_indices_for_initial_vocab = load_models.get_initial_vocab_info(initial_tokenizer)

print('Vocab mask contains '+str(len(initial_vocab))+' entries (written words)') 
print('CMU pronunciation dictionary has '+str(cmu_in_initial_vocab.shape[0])+ ' entries')
print('CMU pronunciation dictionary may have multiple entries for each word')

# confirm yyy treated as a separate character
if 'yyy' not in initial_tokenizer.tokenize('this is a yyy.'):
    raise ValueError('yyy must be treated as a separate character; check the tokenizer')

# build a dataframe of tokens 
# this is slow, because tokenization is slow and single-core
def inflate(row):
    tokens = initial_tokenizer.tokenize(row['gloss_with_punct'])
    return(pd.DataFrame({'token':tokens, 'utterance_id':row['utterance_id']}) )

inflate_path = join(config.prov_csv_dir, 'pvd_utt_glosses_inflated.csv')
if config.regenerate:
    all_tokens = pd.concat([inflate(x) for x in utt_glosses.to_dict('records')])
    all_tokens = all_tokens.merge(utt_glosses)
    all_tokens.to_csv(inflate_path)

else:
    all_tokens = pd.read_csv(inflate_path, na_filter=False)

all_tokens = all_tokens.sort_values(by= ['transcript_id','utterance_order'])

# Assign a token_id (integer in the BERT vocabulary). 
# Because these are from the tokenized utterances, there is no correpsondence with childes-db token ids.
all_tokens['token_id'] = initial_tokenizer.convert_tokens_to_ids(all_tokens['token'])


all_tokens.iloc[0:20]

### Add back IPA, syllable structure, and child ages for child productions

 %%capture
# get the token-level data, esp phonology

save_phono_inflated_path = join(config.prov_csv_dir, 'pvd_utt_glosses_phono_inflated.pkl')

if config.regenerate:

    # get token-level information for Providence
    pvd_chi_tokens = childespy.get_sql_query('select gloss, target_child_name, target_child_age, \
    speaker_code, actual_phonology, model_phonology, transcript_id, utterance_id, \
    token_order from token where speaker_code = "CHI" and corpus_id = '+str(pvd_idx),
        db_version = "2020.1")
    pvd_chi_tokens['gloss'] = [data_cleaning.fix_gloss(x) for x in pvd_chi_tokens.gloss]
    
    # prep the tokens generated from segmenting the utterances
    all_tokens_test = copy.deepcopy(all_tokens) 

    # initialize the fields that need to be populated
    all_tokens_test['actual_phonology'] = ''
    all_tokens_test['model_phonology'] = ''
    all_tokens_test['target_child_age'] = np.nan
    
    # get a set of unique utterances
    _, idx = np.unique(all_tokens_test.utterance_id, return_index=True)
    all_utt_indices = all_tokens_test.utterance_id[np.sort(idx)]
    
    # For fast retrieval of IPA, split pvd_chi_tokens into a dictionary
    pvd_chi_tokens_list = pvd_chi_tokens.groupby(['utterance_id'])
    pvd_chi_tokens_dict = dict(zip(
        [x[0] for x in pvd_chi_tokens_list], 
        [x[1] for x in pvd_chi_tokens_list], 
    ))
    
    # For fast retrival of BERT tokenization
    all_tokens_test_list = all_tokens_test.groupby(['utterance_id'])
    all_tokens_test_dict = dict(zip(
        [x[0] for x in all_tokens_test_list], 
        [x[1] for x in all_tokens_test_list], 
    ))
        
    # Augment the tokens from all_tokens with the IPA from pvd_chi_tokens 
    rvs = [] 
    utts_to_retrieve = yyy_utts.utterance_id.to_list() + success_utts.utterance_id.to_list()
    i=-1
    for utt_index in all_utt_indices: #utts_to_retrieve: #[16760331]:       
        i+=1
        if i % int(len(all_utt_indices) / 100) == 0:
            print(str(np.round((i / (len(all_utt_indices)) * 100),2))+'% complete...')    
            # should learn to use tqdm instead
        if utt_index in utts_to_retrieve:        
            utt_df = copy.deepcopy(all_tokens_test_dict[utt_index])
            utt_df['model_phonology'] = transformers_bert_completions.augment_with_ipa(
              utt_df, pvd_chi_tokens_dict[utt_index],initial_tokenizer, 'model_phonology')
            utt_df['actual_phonology'] = transformers_bert_completions.augment_with_ipa(
              utt_df, pvd_chi_tokens_dict[utt_index],initial_tokenizer, 'actual_phonology')
            utt_df['target_child_age'] = pvd_chi_tokens_dict[utt_index].iloc[0].target_child_age    
            rvs.append(utt_df)  
        else:
            rvs.append(all_tokens_test_dict[utt_index])  
            
    # get the resulting augmented forms back into a dataframe
    all_tokens_phono = pd.concat(rvs)
    
    # add a unique identifier to the BERT tokens
    all_tokens_phono['bert_token_id'] = range(all_tokens_phono.shape[0])
    
    #save the results
    all_tokens_phono.to_pickle(save_phono_inflated_path)
else:
    all_tokens_phono = pd.read_pickle(save_phono_inflated_path)

# Inspect the IPA
all_tokens_phono.loc[all_tokens_phono.actual_phonology != ''][['token','actual_phonology','model_phonology']]

cleaned_inflated_save = join(config.prov_csv_dir, 'pvd_utt_glosses_phono_cleaned_inflated.pkl')

if config.regenerate:    

    # Do the same excludes as were used to identify appropriate utterances
    excludes = ['*','(.)','(..)', '(...)','(....)','(.....)']
    all_tokens_phono.loc[all_tokens_phono.actual_phonology.isin(excludes),'actual_phonology'] =''
    all_tokens_phono.loc[all_tokens_phono.actual_phonology.str.contains('V'),'actual_phonology'] =''
    
    # remap phonology from narrow phonetic transcription to broad phonological transcription
    all_tokens_phono['model_phonology_clean'] = [phonology.phone_remap(x) for x in all_tokens_phono['model_phonology']]
    all_tokens_phono['actual_phonology_clean'] = [phonology.phone_remap(x) for x in all_tokens_phono['actual_phonology']]

    # remove any non-combining diacritical marks
    all_tokens_phono['model_phonology_no_dia'] = [phonology.strip_accents(x) for x in \
    all_tokens_phono['model_phonology_clean']]
    all_tokens_phono['actual_phonology_no_dia'] = [phonology.strip_accents(x) for x in \
    all_tokens_phono['actual_phonology_clean']]
    
    
    # 8/13/21: Changes to limit num vowels on both actual and model phonology.
    # Note will not regen this until later, even if this shows up in near future commits

    all_tokens_phono = data_cleaning.assign_num_vowels_per_phonology(all_tokens_phono, 'actual')
    all_tokens_phono = data_cleaning.assign_num_vowels_per_phonology(all_tokens_phono, 'model')
    
    all_tokens_phono = data_cleaning.combine_num_vowels_phonology(all_tokens_phono)
    
    all_tokens_phono.to_pickle(cleaned_inflated_save)
else:
    all_tokens_phono = pd.read_pickle(cleaned_inflated_save)


all_tokens_phono.loc[all_tokens_phono.actual_phonology_no_dia != '']['actual_phonology_no_dia']

all_tokens_phono.shape

### Identify the tokens that can be evaluated 

# find the tokens in the resulting dataframe that belong to the utterances identified above

initial_vocab


successful_utt_ids = set(success_utts['utterance_id'])

initial_vocab_set = set(initial_vocab)

yyy_utt_ids = set(yyy_utts['utterance_id'])

all_tokens_phono['in_vocab'] = all_tokens_phono['token'].isin(initial_vocab_set)

# 8/1/21: Changed this line to include the vocab constraint.
all_tokens_phono['success_token'] = [(x in successful_utt_ids) and (y) for x, y in 
    zip(all_tokens_phono['utterance_id'], all_tokens_phono['in_vocab'])]
# end changes

all_tokens_phono['yyy_token'] = [x in yyy_utt_ids for x in 
    all_tokens_phono['utterance_id']]


all_tokens_phono.shape

assert '' not in set(all_tokens_phono[all_tokens_phono['num_vowels'] <= 2].actual_phonology)

### Identify the subset of success and failure utterances that have transcriptions

all_tokens_phono['partition'] = 'none'

success_tokens = all_tokens_phono.loc[(all_tokens_phono['success_token']) & 
    (all_tokens_phono['num_vowels'] <= 2) ]
all_tokens_phono.loc[(all_tokens_phono['success_token']) & 
    (all_tokens_phono['num_vowels'] <= 2), 'partition'] = 'success'     
success_tokens.shape

# Successes conditions

success_tokens_check = all_tokens_phono[all_tokens_phono.partition == 'success']
assert all(success_tokens_check['in_vocab'])
assert all(success_tokens_check.utterance_id.isin(successful_utt_ids))
assert all(success_tokens_check['num_vowels'] <= 2)


all_tokens_phono.loc[(all_tokens_phono['success_token']) & 
    (all_tokens_phono['num_vowels'] <= 2)]

yyy_tokens = all_tokens_phono.loc[(all_tokens_phono['yyy_token']) & 
(all_tokens_phono['token'] == 'yyy') & (all_tokens_phono.num_vowels <= 2) ]
all_tokens_phono.loc[(all_tokens_phono['yyy_token']) & 
(all_tokens_phono['token'] == 'yyy') & (all_tokens_phono.num_vowels <= 2),'partition'] = 'yyy'
yyy_tokens.shape

all_tokens_phono.partition.value_counts()

initial_tokenizer.unk_token_id

all_tokens_phono.loc[all_tokens_phono.token == 'xxx','token_id'] = initial_tokenizer.unk_token_id
all_tokens_phono.loc[all_tokens_phono.token == 'yyy','token_id'] = initial_tokenizer.unk_token_id

# this adds the partition information
final_save_path = join(config.prov_csv_dir, 'pvd_utt_glosses_phono_cleaned_inflated_to_next_notebook.pkl')
all_tokens_phono.to_pickle(final_save_path)

# Prevalence of Successes and Failures Across Time

# get number of tokens per age
success_utts['set'] = 'success'
yyy_utts['set'] = 'failure'

# get child age in days associated with each utterance id and plot it

utt_age = chi_phono.groupby('utterance_id').target_child_age.agg(lambda x: np.unique(x)[0]).reset_index()

utts_with_ages = pd.concat([success_utts, yyy_utts]).merge(utt_age)

utts_with_ages['year'] = .5*np.floor(utts_with_ages['target_child_age'] / (365. /2) ) 
print(utts_with_ages.loc[utts_with_ages.set == 'failure'].year.value_counts())
print(utts_with_ages.loc[utts_with_ages.set == 'success'].year.value_counts())

final_utts_save_path = join(config.prov_csv_dir, 'utts_with_ages.csv')
utts_with_ages.to_csv(final_utts_save_path)

assert 'yyy' not in set(success_tokens_check['token'])
# This was the problem that was observed in my iteration of the code.

from datetime import datetime
print(datetime.today())
