==> src/tier_1/5 - Get Non-Providence CHILDES Finetuning Data.ipynb <==
import os
import sys
import pandas as pd
import childespy
import numpy as np
from os.path import join, exists

sys.path.append('../../')
from src.utils import split_gen, data_cleaning, load_splits, configuration, paths
config = configuration.Config()
np.random.seed(config.SEED)

# Generate Train, Val for All and Age splits

#get all of the North American and British English adult and child utterances without xxx or yyy
#concatenate them at the the transcript level
#hold out 20% for validation 

corpora = childespy.get_sql_query('select * from corpus where \
collection_name in ("Eng-NA", "Eng-UK") and data_source = "CHILDES"')
corpora

childes_datasets = ",".join([str(x) for x in corpora.collection_id])

save_path = join(config.finetune_dir, 'utt_glosses.pkl')

if False:
#if config.regenerate:
    utt_glosses = childespy.get_sql_query('select gloss, transcript_id, id, \
    utterance_order, speaker_code, target_child_name, corpus_name, target_child_age, type from utterance where collection_name in ("Eng-NA", "Eng-UK") \
    and collection_id in ('+childes_datasets+') and speaker_code in ("MOT", "FAT","CHI")' , db_version = "2020.1")
    utt_glosses.to_pickle(save_path)
else: 
    utt_glosses = pd.read_pickle(save_path)

utt_glosses = utt_glosses[utt_glosses.gloss != ''] # Added 8/16/21 -- drop empty utterances

all_utt_glosses = utt_glosses.rename(columns = {'id' : 'utterance_id'}).copy()
# end both cites

## "all" split

import imp
imp.reload(paths)
imp.reload(data_cleaning)
imp.reload(split_gen)

# expectations for verbose output.
# 232 should be (4205071, 7) (because you added two extra fields)
# 233 likewise is (3959952, 7)

# Save the partition markings as well, for verification purposes.

success_sample_path = paths.get_sample_csv_path('training', 'Providence', 'all', 'success', age = None, n=5000)

success_sample_path

### No tags

data = data_cleaning.drop_errors(all_utt_glosses)
data_no_tags = data_cleaning.clean_glosses_and_add_speaker_label(data, use_tags = False)
assert not any(['xxx' in s for s in set(data_no_tags.gloss_with_punct)])
assert not any(['yyy' in s for s in set(data_no_tags.gloss_with_punct)])
data_no_tags['tokens'] = [str(x).lower().split(' ') for x in data_no_tags.gloss]

Providence_notags_split_folder = paths.get_directory({
    "task_phase" : 'extract_data',
    "training_split" : 'Providence',
    "training_dataset" : 'all',    
    "test_split" : None,
    "test_dataset" : None,
    "model_type" : None,
    "context_width" : None,
    "n_samples" : config.n_across_time,
    "task_name" : None,
    "use_tags" : False
})
Providence_notags_split_folder

notags_split_glosses_df, notags_train_df = split_gen.exec_split_gen(
    data_no_tags, Providence_notags_split_folder, 'train')

### With tags

data = data_cleaning.drop_errors(all_utt_glosses)
data_with_tags = data_cleaning.clean_glosses_and_add_speaker_label(data, use_tags = True)
assert not any(['xxx' in s for s in set(data_with_tags.gloss_with_punct)])
assert not any(['yyy' in s for s in set(data_with_tags.gloss_with_punct)])
data_with_tags['tokens'] = [str(x).lower().split(' ') for x in data_with_tags.gloss]

Providence_withtags_split_folder = paths.get_directory({
    "task_phase" : 'extract_data',
    "training_split" : 'Providence',
    "training_dataset" : 'all',    
    "test_split" : None,
    "test_dataset" : None,
    "model_type" : None,
    "context_width" : None,
    "n_samples" : config.n_across_time,
    "task_name" : None,
    "use_tags" : True
})
Providence_withtags_split_folder

withtags_split_glosses_df, withtags_train_df = split_gen.exec_split_gen(
    data_with_tags, Providence_withtags_split_folder, 'train')

## "age" split

for use_tags in (True, False):
    
    if use_tags: 
        split_glosses_df = withtags_split_glosses_df
    else:
        split_glosses_df = notags_split_glosses_df
    
    young_df, old_df = split_gen.get_age_split_data(split_glosses_df.copy(), months = config.age_split)

    for args, which_df in zip([('Providence-Age', 'old'), ('Providence-Age', 'young')],
            [old_df, young_df]):

        split, dataset = args
        this_split_folder = paths.get_directory({
                "task_phase" : 'extract_data',
                "training_split" : split,
                "training_dataset" : dataset,    
                "test_split" : None,
                "test_dataset" : None,
                "model_type" : None,
                "context_width" : None,
                "n_samples" : config.n_across_time,
                "task_name" : None,
                "use_tags" : use_tags
            })
        
        split_gen.exec_split_gen(which_df, this_split_folder, 'train')    

from datetime import datetime
print(datetime.today())
