==> src/tier_1/6 - Output Citation vs. Observed Phonology Pairs.ipynb <==
import os
import sys
import rpy2.robjects.lib.ggplot2 as ggplot2
import pandas as pd
import numpy as np
%load_ext rpy2.ipython

sys.path.append('../../')
from src.utils import configuration
config = configuration.Config()

# read in the phono from the pickle so that we have the associated CMU pronunciations
all_tokens_phono = pd.read_pickle(os.path.join(config.prov_csv_dir, 'pvd_utt_glosses_phono_cleaned_inflated_to_next_notebook.pkl'))

excludes = ['', '*','(.)','(..)', '(...)','(....)','(.....)']
phono = all_tokens_phono.loc[~all_tokens_phono.actual_phonology.isin(excludes) & 
                             ~all_tokens_phono.model_phonology.isin(excludes)]
phono.shape

phono.iloc[0]

phono.loc[phono.gloss == 'watch'][['model_phonology_no_dia', 'actual_phonology_no_dia']]

cmu = pd.read_pickle(config.cmu_path)
cmu.head(5)

phono['model_ipa'] = phono.model_phonology_no_dia
phono['actual_ipa'] = [' '.join(list(x)) for x in phono.actual_phonology_no_dia]

'a' in set(list(''.join(phono['actual_ipa'])))

cmu_map = dict(zip(cmu.word, cmu.ipa_short))
def cmu_mapper(x, cmu_map):
    token = x[0]
    model = x[1]
    if x[0] in cmu_map:
        return(cmu_map[x[0]])
    else:
        return(x[1])
    

# cmu inputs need to be changed for the digraphs
phono['cmu'] = [cmu_mapper(x, cmu_map) for x in zip(phono.token, phono.model_ipa)]
phono['cmu'] = [' '.join(list(x.replace('ʧ','tʃ').replace('ʤ','dʒ'))) for x in phono['cmu']]
print(phono.shape)
phono = phono.loc[~phono.cmu.str.contains('\\*')]
print(phono.shape)

phono[['cmu','actual_ipa']]

num_dev = 20000
num_train = 180000
num_test = 20000

for output_dir in ('dev','train','test'):
    dir_to_make = os.path.join(config.project_root, 'output/fst', output_dir)
    if not os.path.exists(dir_to_make):
        os.makedirs(dir_to_make)
        
phono = phono.sample(frac = 1)        

phono[['cmu','actual_ipa']].iloc[0:num_dev].to_csv(os.path.join(
    config.project_root, 'output/fst/dev/chi_dev.tsv'),
    sep='\t', header=False, index=False)
phono[['cmu','actual_ipa']].iloc[num_dev:num_dev+num_train].to_csv(os.path.join(
    config.project_root, 'output/fst/train/chi_train.tsv'),
    sep='\t', header=False, index=False)
phono[['cmu','actual_ipa']].iloc[num_dev+num_train:num_dev+num_train+num_test].to_csv(os.path.join(
    config.project_root,'output/fst/test/chi_test.tsv'),
    sep='\t', header=False, index=False)

phono[['model_ipa_split']] = [' '.join(list(x)) for x in phono['cmu']]
phono[['model_ipa_split']].iloc[num_dev:num_dev+num_train].to_csv(os.path.join(config.project_root,
    'output/fst/train/chi_inputs.tsv'),
    sep='\t', header=False, index=False)

# Split the Phonology Training Data by Child

# what we want now is dev / train / test for each of the children

dev_tokens = phono.iloc[0:num_dev]
training_tokens = phono.iloc[num_dev:num_dev+num_train]
test_tokens = phono.iloc[num_dev+num_train:num_dev+num_train+num_test]

training_tokens_grouped = training_tokens.groupby(training_tokens['target_child_name'])
dev_tokens_grouped = dev_tokens.groupby(dev_tokens['target_child_name'])
test_tokens_grouped = test_tokens.groupby(test_tokens['target_child_name'])

for child in training_tokens_grouped.groups.keys():    
    
    dev_df = dev_tokens_grouped.get_group(child)    
    dev_df[['cmu','actual_ipa']].to_csv(os.path.join(config.project_root,'output/fst/dev/'+child+'_dev.tsv'), sep='\t', header=False, index=False)
    
    train_df = training_tokens_grouped.get_group(child)
    train_df[['cmu','actual_ipa']].to_csv(os.path.join(config.project_root,'output/fst/train/'+child+'_train.tsv'), sep='\t', header=False, index=False)
    
    test_df  = test_tokens_grouped.get_group(child)
    test_df[['cmu','actual_ipa']].to_csv(os.path.join(config.project_root,'output/fst/test/'+child+'_test.tsv'), sep='\t', header=False, index=False)

# Make a UTF-8 symbol set that covers all of the characters across all children, train + test sets

all_sounds = [x.split(' ') for x in  phono['actual_ipa']]
phones = np.unique([item for subl in all_sounds for item in subl])
phones

# cmu
all_cmu_sounds = [x.split(' ') for x in  phono['cmu']]
cmu_phones = np.unique([item for subl in all_cmu_sounds for item in subl])
cmu_phones

set(cmu_phones) - set(phones)

# check what was loaded by likelihoods and see if it is outside of this set. Actual_ipa may be from the CMU renditions, which are different

'ɲ' in phones

utf8_sym = pd.concat([
    pd.DataFrame({'utf8':['<epsilon>'], 'sym': [0]}),
    pd.DataFrame({'utf8':phones, 'sym': [ord(x) for x in phones]}),
])
utf8_sym

utf8_sym.to_csv('../../src/external/all_child_phones.sym', header=False, index=False, sep='\t')
