==> src/tier_1/9 - Get Switchboard Finetuning Data-Copy1.ipynb <==
import os
import sys
import pandas as pd
import numpy as np
import re
import glob
from os.path import join, exists

sys.path.append('../../')
from src.utils import split_gen, data_cleaning, load_splits, configuration, paths
config = configuration.Config()
np.random.seed(config.SEED)

# Download and Preprocess Switchboard

!mkdir ../../data/switchboard

!curl -o ../../data/switchboard/ptree_word_alignments.tar.gz https://isip.piconepress.com/projects/switchboard/releases/ptree_word_alignments.tar.gz

%%capture
!tar -xvzf ../../data/switchboard/ptree_word_alignments.tar.gz -C ../../data/switchboard/;

def clean_swbd_word(w):
    cleaned_w = re.sub("\\[laughter-", "", w)
    cleaned_w = re.sub("\\[vocalized", "", cleaned_w)
    cleaned_w = re.sub("\\[", "", cleaned_w)
    cleaned_w = re.sub("\\]", "", cleaned_w)
    cleaned_w = re.sub("\\-$", "", cleaned_w)
    cleaned_w = re.sub("^\\-", "", cleaned_w)
    cleaned_w = re.sub("\\{", "", cleaned_w)
    cleaned_w = re.sub("\\}", "",cleaned_w)        
    cleaned_w = re.sub("_1", "",cleaned_w)        
    cleaned_w = re.sub("^-$", "",cleaned_w)        
    cleaned_w = re.sub("^.*/$", "",cleaned_w)        
    cleaned_w = cleaned_w.lower()
    return(cleaned_w)
    

def read_penn_tagged_swithboard_file(path, tokens_to_exclude):
    df = pd.read_table(path, sep='\t', header=None)
    df.columns = ['id', 'speaker.utt','start','stop', 'unk','word1','word2']
    
    df = df.loc[~df.word2.isin(tokens_to_exclude)]
    df['utt_id'] = [int(x.split('.')[1]) for x in df['speaker.utt']]
    df['speaker'] = [x.split('.')[0] for x in df['speaker.utt']]
    df['conversation'] = [re.sub('A|B','',  x.split('-')[0]) for x  in df['id']]

    
    # preprocess the words with the function below
    df['cleaned_word'] = [clean_swbd_word(w) for w in df['word2']]
    
    # pull out the start time for everything with the new system
    utts = df.groupby(['utt_id', 'speaker', 'conversation']).cleaned_word.agg(lambda x: ' '.join(x)).reset_index()
    return(utts)
                              

def read_penn_tagged_swbd_pair(a_path, tokens_to_exclude):
    b_path = a_path.replace('A-','B-')
    
    a_df = read_penn_tagged_swithboard_file(a_path, tokens_to_exclude)
    b_df = read_penn_tagged_swithboard_file(b_path, tokens_to_exclude)
        
    combined = pd.concat([a_df, b_df]).sort_values(by=['utt_id'])
    return(combined)


tokens_to_exclude = ["[vocalized-noise]", "[noise]", "[laughter]",
"[silence]", "","+++"]

a_paths = glob.glob('../../data/switchboard/data/alignments/*/*A-*.text')

from joblib import Parallel, delayed
results = Parallel(n_jobs=48)(delayed(read_penn_tagged_swbd_pair)(a_path,tokens_to_exclude) for a_path in a_paths)

all_convos = pd.concat(results)

all_convos

swbd_txt_path = os.path.join(config.project_root, 'output/finetune/switchboard/all', 'switchboard_cleaned.txt')
all_convos[['cleaned_word']].to_csv(swbd_txt_path, header=False, index=False)

all_convos.columns =  ['utt_id','speaker','conversation','gloss']
all_convos['gloss_with_punct'] = [x+'.' for x in all_convos['gloss']]
all_convos['utterance_id'] = range(all_convos.shape[0])

all_convos['transcript_id'] = all_convos['conversation']

## "all" split

switchboard_split_folder = paths.get_directory({
            "task_phase" : 'extract_data',
            "training_split" : 'Switchboard',
            "training_dataset" : 'all',    
            "test_split" : None,
            "test_dataset" : None,
            "model_type" : None,
            "context_width" : None,
            "n_samples" : config.n_across_time,
            "task_name" : None,
            "use_tags" : False
})

split_glosses_df, train_df = split_gen.exec_split_gen(all_convos, switchboard_split_folder, 'train') 

all_convos['word_count'] = [len(x.split(' ')) for x in all_convos['gloss']]

np.sum(all_convos['word_count'])
# only 1.5m words -- is this after removing silence tokens, etc.?
